requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
chromadb>=0.4.0
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
streamlit>=1.28.0
requests>=2.31.0
plotly>=5.17.0
python-multipart>=0.0.6